+++ frontend/src/components/ItemMarket.vue
@@ -47,12 +47,12 @@
 
 <script>
 import axios from "axios"
-import mixins from "@/mixins/PixyShip.vue.js"
-import itemMixins from "@/mixins/Item.vue.js"
+import PixyShipMixin from "@/mixins/PixyShip.vue.js"
+import ItemMixin from "@/mixins/Item.vue.js"
 import _ from 'lodash'
 
 export default {
-  mixins: [mixins, itemMixins],
+  mixins: [PixyShipMixin, ItemMixin],
 
   props: {
     item: null,