+++ CHANGELOG.md
@@ -5,13 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.6] - 2021-11-24
+
+### Added
+
+- Items page: Type and Subtype columns sortable
+- Items page: Display training % for instant training item
+
+## [2.2.5] - 2021-11-14
+
+### Added
+
+- Ships page: extended filter
+- Save sorting in URL
+
 ## [2.2.4] - 2021-11-12
 
 ### Changed
 
 - Ships page: display full repair time
 
-### ixed
+### Fixed
 
 - Rooms page: wrong capacity label and value
 
@@ -198,6 +212,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Last Jeff commit
 
 [unreleased]: https://github.com/solevis/pixyship/compare/main...develop
+[2.2.6]: https://github.com/solevis/pixyship/compare/2.2.5...2.2.6
+[2.2.5]: https://github.com/solevis/pixyship/compare/2.2.4...2.2.5
 [2.2.4]: https://github.com/solevis/pixyship/compare/2.2.3...2.2.4
 [2.2.3]: https://github.com/solevis/pixyship/compare/2.2.2...2.2.3
 [2.2.2]: https://github.com/solevis/pixyship/compare/2.2.1...2.2.2